Refactor code and update db schema

+++ src/main/java/com/microservices/projectservice/service/FormService.java
@@ -1,99 +1,67 @@
 package com.microservices.projectservice.service;
 
 import com.microservices.projectservice.dto.request.FormCreateRequest;
-import com.microservices.projectservice.dto.response.FormResponse;
 import com.microservices.projectservice.dto.request.FormUpdateRequest;
-import com.microservices.projectservice.dto.response.PagingObjectsResponse;
 import com.microservices.projectservice.entity.Form;
-import com.microservices.projectservice.entity.Project;
-import com.microservices.projectservice.entity.Stage;
 import com.microservices.projectservice.exception.IllegalAttributeException;
 import com.microservices.projectservice.exception.NoEntityFoundException;
 import com.microservices.projectservice.repository.FormRepository;
 import com.microservices.projectservice.repository.ProjectRepository;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.validation.annotation.Validated;
+
+import java.util.Collections;
+import java.util.Objects;
 
 @Service
-@Validated
 @RequiredArgsConstructor
 public class FormService {
 
     private final FormRepository formRepository;
     private final ProjectRepository projectRepository;
+    private final ProjectService projectService;
 
-    public PagingObjectsResponse<FormResponse> getAllForms(
-            @NotBlank(message = "Project ID cannot be null/blank when getting all forms.") String projectId,
-            @Min(value = 0, message = "Invalid page number (must positive) when getting all forms.")
-            @NotNull(message = "Page number cannot be null when getting all forms.")
-            Integer pageNumber,
-            @Min(value = 1, message = "Invalid page size (must greater than 0) when getting all forms.")
-            @NotNull(message = "Page size cannot be null when getting all forms.")
-            Integer pageSize
-    ) throws NoEntityFoundException {
+    public Page<Form> getAllForms(String projectId,
+                                  Integer pageNumber,
+                                  Integer pageSize) throws NoEntityFoundException {
         if (!projectRepository.existsById(projectId))
             throw new NoEntityFoundException("No project found with id: " + projectId);
 
         var pageable = PageRequest.of(pageNumber, pageSize, Sort.by("createdAt").descending());
-        var forms = formRepository.findAllByProjectOwner_Id(projectId, pageable)
-                .map(this::mapToResponse);
-        return new PagingObjectsResponse<>(
-                forms.getTotalPages(),
-                forms.getTotalElements(),
-                forms.getNumber(),
-                forms.getSize(),
-                forms.getNumberOfElements(),
-                forms.isFirst(),
-                forms.isLast(),
-                forms.toList()
-        );
+        return formRepository.findAllByProjectOwner_Id(projectId, pageable);
     }
 
-    public FormResponse getForm(
-            @NotBlank(message = "Form ID cannot be null/blank when getting a form.") String formId
-    ) throws NoEntityFoundException {
-        var form = findForm(formId);
-        return mapToResponse(form);
+    public Form getForm(String formId) throws NoEntityFoundException {
+        return formRepository.findById(formId)
+                .orElseThrow(() -> new NoEntityFoundException("No form found with id: " + formId));
     }
 
-    public String createForm(
-            @NotNull(message = "Creating form data cannot be null.")
-            @Valid
-            FormCreateRequest formCreateRequest
-    ) throws NoEntityFoundException {
-        var project = findProject(formCreateRequest.projectOwnerId());
+    public String createForm(FormCreateRequest body) throws NoEntityFoundException {
+        var project = projectService.getProject(body.projectOwnerId());
         var form = Form.builder()
-                .title(formCreateRequest.title())
-                .description(formCreateRequest.description())
+                .title(body.title())
+                .description(body.description())
                 .projectOwner(project)
                 .build();
         return formRepository.save(form).getId();
     }
 
-    public void updateForm(
-            @NotBlank(message = "Form ID cannot be null/blank when updating a form.") String formId,
-            @NotNull(message = "Updating form data cannot be null.")
-            @Valid
-            FormUpdateRequest formUpdateRequest
-    ) throws IllegalAttributeException, NoEntityFoundException {
+    public void updateForm(String formId, FormUpdateRequest body)
+            throws IllegalAttributeException, NoEntityFoundException {
         var isUpdate = false;
-        var form = findForm(formId);
+        var form = getForm(formId);
 
-        var title = formUpdateRequest.title();
+        var title = body.title();
         if (title != null) {
-            if (title.isBlank()) throw new IllegalAttributeException("Form title cannot be blank.");
+            if (title.isBlank()) throw new IllegalAttributeException("title cannot be blank.");
             form.setTitle(title);
             isUpdate = true;
         }
 
-        var description = formUpdateRequest.description();
+        var description = body.description();
         if (description != null) {
             form.setDescription(description);
             isUpdate = true;
@@ -102,33 +70,12 @@ public class FormService {
         if (isUpdate) formRepository.save(form);
     }
 
-    public void deleteForm(
-            @NotBlank(message = "Form ID cannot be null/blank when updating a form.") String formId
-    ) throws NoEntityFoundException {
-        var form = findForm(formId);
+    public void deleteForm(String formId) throws NoEntityFoundException, IllegalAttributeException {
+        var form = getForm(formId);
+        var usageStages = Objects.requireNonNullElse(form.getUsageStages(), Collections.emptyList());
+        if (!usageStages.isEmpty()) throw new IllegalAttributeException(
+                "The form cannot be deleted. Because, currently, it is being used by several stages.");
         formRepository.delete(form);
-    }
-
-    private Form findForm(String formId) throws NoEntityFoundException {
-        return formRepository.findById(formId)
-                .orElseThrow(() -> new NoEntityFoundException("No form found with id: " + formId));
-    }
-
-    private Project findProject(String projectId) throws NoEntityFoundException {
-        return projectRepository.findById(projectId)
-                .orElseThrow(() -> new NoEntityFoundException("No project found with id: " + projectId));
-    }
-
-    private FormResponse mapToResponse(Form form) {
-        var usageStageIds = form.getUsageStages().stream().map(Stage::getId).toList();
-        return new FormResponse(
-                form.getId(),
-                form.getTitle(),
-                form.getDescription(),
-                form.getProjectOwner().getId(),
-                form.getCreatedAt().getTime(),
-                usageStageIds
-        );
     }
 
 }
